refactor: Rename to image
